Adding and moving unit and snapshot tests #689
Previously, all tests for searching existed in searchButton component. These have now been moved to the parent component. Snapshots have been re-rendered and tests have also been made for searchBoxContainerSide to increase coverage.
