+++ packages/datagateway-search/src/searchBoxContainer.component.test.tsx
@@ -1,43 +1,48 @@
 import React from 'react';
 import { ReactWrapper } from 'enzyme';
 
-import thunk from 'redux-thunk';
-import configureStore from 'redux-mock-store';
-import { StateType } from './state/app.types';
-import { initialState as dgSearchInitialState } from './state/reducers/dgsearch.reducer';
-import { dGCommonInitialState } from 'datagateway-common';
-
 import { createShallow } from '@material-ui/core/test-utils';
-import { MemoryRouter } from 'react-router';
 import SearchBoxContainer from './searchBoxContainer.component';
+import SearchBoxContainerSide from './searchBoxContainerSide.component';
 
 jest.mock('loglevel');
 
 describe('SearchBoxContainer - Tests', () => {
   let shallow;
-  let state: StateType;
+
+  const testInitiateSearch = jest.fn();
 
   const createWrapper = (path: string): ReactWrapper => {
-    const mockStore = configureStore([thunk]);
+    return shallow(<SearchBoxContainer initiateSearch={testInitiateSearch} />);
+  };
+
+  beforeEach(() => {
+    shallow = createShallow({ untilSelector: 'Grid' });
+  });
+
+  it('renders searchBoxContainer correctly', () => {
+    const wrapper = createWrapper('/');
+
+    expect(wrapper).toMatchSnapshot();
+  });
+});
+
+describe('SearchBoxContainerSide - Tests', () => {
+  let shallow;
+
+  const testInitiateSearch = jest.fn();
+
+  const createWrapper = (path: string): ReactWrapper => {
     return shallow(
-      <MemoryRouter initialEntries={[{ key: 'testKey', pathname: path }]}>
-        <SearchBoxContainer store={mockStore(state)} />
-      </MemoryRouter>
+      <SearchBoxContainerSide initiateSearch={testInitiateSearch} />
     );
   };
 
   beforeEach(() => {
     shallow = createShallow({ untilSelector: 'Grid' });
-
-    state = JSON.parse(
-      JSON.stringify({
-        dgcommon: dGCommonInitialState,
-        dgsearch: dgSearchInitialState,
-      })
-    );
   });
 
-  it('renders searchBoxContainer correctly', () => {
+  it('renders searchBoxContainerSide correctly', () => {
     const wrapper = createWrapper('/');
 
     expect(wrapper).toMatchSnapshot();
